requirements.txt
# Core ML
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
scipy==1.12.0

# NLP
nltk==3.8.1
spacy==3.7.4

# MLflow
mlflow==2.12.1

# AWS
boto3==1.34.113
botocore==1.34.113
s3fs==2024.3.1

# API
fastapi==0.110.2
uvicorn==0.29.0
pydantic==2.7.1

# Docker SDK
docker==7.0.0

# Kubernetes (EKS)
kubernetes==29.0.0

# Config & YAML
PyYAML==6.0.1
python-dotenv==1.0.1

# Logging & Utils
loguru==0.7.2
tqdm==4.66.4

# Testing
pytest==8.2.1
pytest-cov==5.0.0

# Linting & formatting
black==24.4.2
flake8==7.0.0
isort==5.13.2

# Gunicorn for production API
gunicorn==21.2.0